requirements.txt
gymnasium==0.27.1
matplotlib==3.6.3
numpy==1.24.1
opencv_python==4.7.0.68
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
autorom==0.4.2
